Add ItemData abstraction layer
ItemData is not tied to the existence of a GameObject, allowing for
  delayed instantiation after deserializing the game state, which is
  exactly what this is doing now
Fix held or attached items not being preserved through game state
  serialization -> deserialization cycles
Add ItemData.lua file to generate the ItemData.cs file. The generation
  script is currently not linkable nor included in this repo, not sure
  how I'm going to handle that. Should probably just put it on github
  or make a github gist
Add vpmignore. Unteated, I hope the glob matcher that VPMPublish is
  using works with it, I don't remember. Been a while
Add .vscode folder to .gitignore